requirements.txt
speechrecognition
openai
elevenlabs
pyaudio
pyinstaller
python-dotenv